feat: use generic error for clearer message and easier trace, no need backtrace
